% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_discography.R
\name{get_discography}
\alias{get_discography}
\title{Get Discography}
\usage{
get_discography(name)
}
\arguments{
\item{name}{Character (length = 1).
An artist's name}
}
\value{
Returns a data frame with the artists, song names,
Genius ID, Genius URL, primary artist, primary artist Genius ID,
and primary artist Genius URL
}
\description{
Gets full information for an artist
}
\examples{
\dontrun{
# Search for "The Beatles"
get_discography("The Beatles")}

}
\author{
Alberto Almuina and Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
