% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lyrics.R
\name{get_lyrics}
\alias{get_lyrics}
\title{Get Song Lyrics}
\usage{
get_lyrics(artist_name, song_names, urls = NULL, verbose = TRUE)
}
\arguments{
\item{artist_name}{Character (length = 1).
Uses the artist name to get correct lyrics.
The "primary artist" will be used to identify the
appropriate artist}

\item{song_names}{Character (vector).
Name or names of songs to get lyrics for (narrows down by \code{artist_name})}

\item{urls}{Character (vector).
URLs for the songs (rather than using artist and song names).
Skips search for proper information and provides exactly the URL
to the lyrics desired}

\item{verbose}{Boolean (length = 1).
Whether progress on which song lyrics are being obtained
should be printed.
Defaults to \code{TRUE}}
}
\value{
Returns a data frame with artist, song name,
part of the song (e.g., verse, chorus), and lyric for each part
}
\description{
Get song lyrics and parts from a song or many songs by
an artist
}
\examples{
\dontrun{
get_lyrics(
  artist_name = "The Beatles",
  song_names = c(
    "Lucy in the Sky with Diamonds",
    "Magical Mystery Tour"
  )
)}

}
\author{
Alberto Almuina and Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
