% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_information.R
\name{search_information}
\alias{search_information}
\title{Search By Artist or Song for Information}
\usage{
search_information(name, type = c("artist", "song"))
}
\arguments{
\item{name}{Character (length = 1).
Either an artist or a song}

\item{type}{Character (length = 1).
Specify whether \code{name} is an artist or a song}
}
\value{
Returns a data frame with the artists, song names,
Genius ID, Genius URL, primary artist, primary artist Genius ID,
and primary artist Genius URL
}
\description{
Get basic Genius information for an artist or song title
}
\examples{
\dontrun{
# Search for "The Beatles"
search_information("The Beatles", type = "artist")}

}
\author{
Alberto Almuina and Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
